+++ src/file_entry.c
@@ -169,7 +169,7 @@ fpart_hook(const char *cmd, const struct program_options *options,
 
         if(options->verbose >= OPT_VERBOSE)
             fprintf(stderr, "Executing pre-part #%ju hook: '%s'\n",
-                *live_partition_index, cmd);
+                display_index(*live_partition_index, options), cmd);
 
         /* FPART_HOOKTYPE (pre-part) */
         malloc_size = strlen(env_fpart_hooktype_name) + 1 +
@@ -192,7 +192,7 @@ fpart_hook(const char *cmd, const struct program_options *options,
 
         if(options->verbose >= OPT_VERBOSE)
             fprintf(stderr, "Executing post-part #%ju hook: '%s'\n",
-                *live_partition_index, cmd);
+                display_index(*live_partition_index, options), cmd);
 
         /* FPART_HOOKTYPE (post-part) */
         malloc_size = strlen(env_fpart_hooktype_name) + 1 +
@@ -228,13 +228,13 @@ fpart_hook(const char *cmd, const struct program_options *options,
     /* FPART_PARTNUMBER */
     if(live_partition_index != NULL) {
         malloc_size = strlen(env_fpart_partnumber_name) + 1 +
-            get_num_digits(*live_partition_index) + 1;
+            get_num_digits(display_index(*live_partition_index, options)) + 1;
         if_not_malloc(env_fpart_partnumber_string, malloc_size,
             retval = 1;
             goto cleanup;
         )
         snprintf(env_fpart_partnumber_string, malloc_size, "%s=%ju",
-            env_fpart_partnumber_name, *live_partition_index);
+            env_fpart_partnumber_name, display_index(*live_partition_index, options));
         if(push_env(env_fpart_partnumber_string, &envp) != 0) {
             retval = 1;
             goto cleanup;
@@ -404,12 +404,12 @@ live_print_file_entry(char *path, fsize_t size,
         if(out_template != NULL) {
             /* compute live_status.filename "out_template.i\0" */
             size_t malloc_size = strlen(out_template) + 1 +
-                get_num_digits(live_status.partition_index) + 1;
+                get_num_digits(display_index(live_status.partition_index, options)) + 1;
             if_not_malloc(live_status.filename, malloc_size,
                 return (1);
             )
             snprintf(live_status.filename, malloc_size, "%s.%ju", out_template,
-                live_status.partition_index);
+                display_index(live_status.partition_index, options));
         }
 
         /* execute pre-partition hook */
@@ -440,7 +440,8 @@ live_print_file_entry(char *path, fsize_t size,
 
     if(out_template == NULL) {
         /* no template provided, just print to stdout */
-        fprintf(stdout, "%ju (%ju): %s\n", live_status.partition_index, size,
+        fprintf(stdout, "%ju (%ju): %s\n",
+            display_index(live_status.partition_index, options), size,
             path);
     }
     else {
@@ -467,7 +468,8 @@ live_print_file_entry(char *path, fsize_t size,
         /* display added partition */
         if(options->verbose >= OPT_VERBOSE)
             fprintf(stderr, "Filled part #%ju: size = %ju, %ju file(s)\n",
-                live_status.partition_index, live_status.partition_size,
+                display_index(live_status.partition_index, options),
+                live_status.partition_size,
                 live_status.partition_num_files);
 
         /* close fd or flush buffer */
@@ -919,7 +921,8 @@ uninit_file_entries(struct file_entry *head, struct program_options *options)
         if((options->verbose >= OPT_VERBOSE) &&
             (live_status.partition_num_files > 0))
             fprintf(stderr, "Filled part #%ju: size = %ju, %ju file(s)\n",
-                live_status.partition_index, live_status.partition_size,
+                display_index(live_status.partition_index, options),
+                live_status.partition_size,
                 live_status.partition_num_files);
 
         /* flush buffer or close last file if necessary */
@@ -967,7 +970,8 @@ print_file_entries(struct file_entry *head, pnum_t num_parts,
     /* no template provided, just print to stdout and return */
     if(out_template == NULL) {
         while(head != NULL) {
-            fprintf(stdout, "%ju (%ju): %s\n", head->partition_index,
+            fprintf(stdout, "%ju (%ju): %s\n",
+                display_index(head->partition_index, options),
                 head->size, head->path);
             head = head->nextp;
         }