+++ src/partition.h
@@ -47,6 +47,6 @@ int add_partitions(struct partition **head, pnum_t num_parts,
 void uninit_partitions(struct partition *head);
 pnum_t find_smallest_partition_index(struct partition *head);
 struct partition * get_partition_at(struct partition *head, pnum_t index);
-void print_partitions(struct partition *head);
+void print_partitions(struct partition *head, struct program_options *options);
 
 #endif /* _PARTITION_H */